t2/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.10
langchain-community>=0.0.16
langchain-groq>=0.1.0
langchain-pinecone>=0.1.0
pinecone-client>=3.0.0
groq>=0.4.0
sentence-transformers>=2.2.2
torch>=2.0.0
pymupdf>=1.23.0
pillow>=10.0.0
python-dotenv>=1.0.0
rich>=13.0.0
pymupdf>=1.23.0
pillow>=10.0.0
python-dotenv>=1.0.0
rich>=13.0.0